Enforce structured projection identity
